+++ frontend/components/lv4/WorkDetail.tsx
@@ -1,10 +1,9 @@
-import { FC, useState, useEffect } from "react";
-import { Work, Original } from "model";
+import { FC } from "react";
+import { Work } from "model";
 import { WorkImage } from "components/lv1/WorkImage";
 import styled from "styled-components";
 import Constants from "styles/Constants";
-import { firestore } from "lib/firebase/client";
-import { compactMap } from "helper/array";
+import { useWorkOriginals } from "hooks/work/useWorkOriginals";
 
 type Props = {
   work: Work
@@ -14,28 +13,6 @@ const WorkTitle = styled.div`
   font-size: ${Constants.FONT.LARGE}px;
   font-weight: ${Constants.FONT_WEIGHT.BOLD};
 `
-
-const useWorkOriginals = (work: Work): { loading: boolean, originals:Original[] } => {
-  const [originals, setOriginals] = useState<Original[]>([])
-  const [loading, setLoading] = useState(true)
-
-  useEffect(() => {
-    let cancel = false
-    const retrive = async () => {
-      setLoading(true)
-      const result = await firestore.collection('works').doc(work.id).collection('originals').limit(10).get()
-      if (!cancel) {
-        const originals = compactMap(result.docs, doc => new Original(doc.data()))
-        setOriginals(originals)
-        setLoading(false)
-      }
-    }
-    retrive()
-    return () => { cancel = true }
-  }, [])
-
-  return { loading, originals }
-}
 
 const WorkOriginal: FC<{ work: Work }> = (props) => {
   const { loading, originals } = useWorkOriginals(props.work)
@@ -50,7 +27,7 @@ const WorkOriginal: FC<{ work: Work }> = (props) => {
   return (
     <div>
       {originals.map((original, i) => (
-        <div key={`original_${i}`}>{original.data}</div>
+        <div key={`original_${i}`}>{JSON.stringify(original)}</div>
       ))}
     </div>
   )
